University almost done

+++ src/main/java/model/ComputerScienceStudent.java
@@ -2,7 +2,7 @@ package model;
 
 public class ComputerScienceStudent extends StudentAbstract{
 
-    public ComputerScienceStudent(String lastName, String firstName,  int id, String gender, int age, String majorCourse, double studentGrade, boolean representative){
+    public ComputerScienceStudent(String lastName, String firstName,  int id, String gender, int age, String majorCourse, float studentGrade, boolean representative){
         super(lastName, firstName, id, gender, age, majorCourse, studentGrade, representative);
     }
