+++ src/main/java/model/HistoryStudent.java
@@ -4,8 +4,9 @@ package model;
 
 public class HistoryStudent extends StudentAbstract{
 
-    public HistoryStudent(String lastName, String firstName,  int id, String gender, int age, String majorCourse, double studentGrade, boolean representative){
+    public HistoryStudent(String lastName, String firstName,  int id, String gender, int age, String majorCourse, float studentGrade, boolean representative){
         super(lastName, firstName, id, gender, age, majorCourse, studentGrade, representative);
+
     }
 
     @Override
